Added Dashboard pages, moved components to designated folder
